Make the sub resource load also skip the fetch handler if skippable.
In the previous implementation, we did not skip service worker for
fetching sub resource load.  Since we do not start the service worker
when the fetch handler is decided to be skipped, sub resource load may
get stuck.

This CL make the renderer behave as if the fetch handler does not
exist if the fetch handler is skipped in the main resource load.
To let the renderer know the fetch handler should be skipped, we
introduced the effective fetch handler type in
ControllerServiceWorkerInfo.  If the effective fetch handler type
is skippable, the renderer returns nullptr when the controller is
requested.

You may wonder why we need to introduce the effective fetch handler
type in addition to the existing fetch_handler_type.  That is because
we cannot change the fetch_handler_type behavior for the service
worker registration and the metrics.  Since the service worker
registration is persistent data, I do not think it is good idea to
change its contents by the flag.  For metrics, we want to compare the
same fetch handler case with the different flags.  The
fetch_handler_type should also need to be persistent here.
Note that FCP/LCP with skippable fetch handler type is taken in this
way.

Bug: 1364999
Bug: 1347319
Change-Id: I29f5af4589d5a675bf033dd74a17c975dd1a7722
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3905774
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Reviewed-by: Shunya Shishido <sisidovski@chromium.org>
Reviewed-by: Minoru Chikamune <chikamune@chromium.org>
Commit-Queue: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Reviewed-by: Sam McNally <sammc@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1052295}
NOKEYCHECK=True
GitOrigin-RevId: f9a238364dead2a4647b08b228acf7d28af20661