Chrome privileged helper: copy updater to temp and verify before run.
This is an SMJobBlessed helper that tries to convert Chrome to a system-
wide (root-owned) installation.

The intent is to prevent the helper from running arbitrary code as root.

The assumption is that /tmp is a secure place for root to use temp dirs.

Bug: 1361381
Change-Id: I0bec05fb0f3ed89d32c971c658119b4dd6345343
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3885176
Commit-Queue: Joshua Pawlicki <waffles@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1052353}
NOKEYCHECK=True
GitOrigin-RevId: 6bf482d6ce3acabe353eab64af99eef5963cdf1d