Add resize-visual, overlays-content to virtual-keyboard
Adds resize-visual and overlays-content as options to the
virtual-keyboard option in the viewport meta tag. This meta tag
option is guarded by the currently unshipped OSKResizesVisualViewport
flag.

`resize-visual` is added for completeness but will be a no-op once the
flag ships. `overlays-content` allows setting the existing imperative
API declaratively, resolving 1250883.

With this change, the VirtualKeyboardEnum now has values for all modes
as well as "UNSET". CompositorViewHolder no longer treats UNSET as a
mode; instead turning UNSET into whichever real mode is the default,
based on the flag. This makes the code easier to understand and
explicitly test.

Only tangentially related: Split a CompositorViewHolderUnitTest that was
failing with the OSKResizesVisualViewport flag into two tests with
expectations depending on which mode is used.

Bug: 1353728, 1250883
Change-Id: I09fa556608e2da05266a5b437b5bc2c2ac0f1b3b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3914523
Auto-Submit: David Bokan <bokan@chromium.org>
Commit-Queue: David Bokan <bokan@chromium.org>
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1052554}
NOKEYCHECK=True
GitOrigin-RevId: 1c3cb248e5dd64cc6a719bbe28fb7294e7e76bfa