Roll FFMpeg, including fixes
Rolls src/third_party/ffmpeg:
3dd1a90db => b9f01c3c5

$ git log 3dd1a90db..b9f01c3c5 --date=short --no-merges --format='%ad %ae %s'
2022-09-22 tmathmeyer@gmail.com README.chromium file
2022-09-22 tmathmeyer@gmail.com GN Configuration
2022-09-20 jamrial@gmail.com x86/aacpsdsp: add ps_hybrid_analysis_fma3
2022-09-20 jamrial@gmail.com x86/aacpsdsp: precompute constant factors
2022-09-20 remi@remlab.net lavc/aacpsdsp: precompute constant factors
(...)
2022-09-10 dev@lynne.ee lavu/tx: generalize MDCTs
2022-09-10 dev@lynne.ee lavu/tx: add the inplace flag to PFA FFTs
2022-09-10 dev@lynne.ee lavu/tx: propagate the codelet flags into the context
2022-09-07 andreas.rheinhardt@outlook.com fate/matroska: Add test for updating AV1 extradata
2022-09-07 andreas.rheinhardt@outlook.com tests/fate-run: Allow to set input options for encoding pass

the big change this time in ffmpeg is that the vorbis decoder now uses
pts to determine if a frame should be discarded post-seeking, rather
than always discarding the first frame after flushing. The commit can
be found here:
https://github.com/FFmpeg/FFmpeg/commit/8fc2dedfe6e

This didn't actually break any of chrome's behavior, but it did break
a bunch of tests, since we had never attached the pts to our testing
DecoderBuffer blobs.

For AudioDecoderBrokerTest, I had to add an option to get a
DecoderBuffer with an assigned timestamp from test_data_util.cc.

For pipeline_integration_test.cc, The audio hash changes, as we now
are keeping around one more packet post-seek, which I verified is the
correct behavior by comparing it to ffplay.

For the video-seek-past-end-paused test, this change actually exposed
that the test itself was racy because |video.ended| isn't actually
guaranteed to be set to true by the time that |video.onpause| is
triggered. The presence of one more packet in the audio stream now
makes it delayed by _just_ enough time to flake with about an 80%
failure rate when played locally, and flake on a few of the trybots as
well. Since its not actually part of the spec that this ended state
is reached, I've just removed it from the assertion set.

The other changes to tests have already landed as they were more
general and didn't require rolling the ffmpeg deps forward.

Bug: 1366548
Change-Id: Ifaab533796ef11703e76ce22982121bfb5899dbd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3915268
Commit-Queue: Matthew Wolenetz <wolenetz@chromium.org>
Reviewed-by: Matthew Wolenetz <wolenetz@chromium.org>
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Auto-Submit: Ted (Chromium) Meyer <tmathmeyer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1052557}
NOKEYCHECK=True
GitOrigin-RevId: d42c7e5f8b8289c22d8daae4d76472435303f7d6