[last-baseline] Propagate last baseline from grid layout.
This modifies grid's baseline accumulator to support three(!) potential
baselines for the first & last baseline.

Firstly we prefer the major/minor baseline for the first/last baseline
of the first/last row which has content (if any).

Then we prefer the minor/major baseline for the first/last baseline of
the first/last row which has content (if any).

Finally fallback baseline is the first/last item in the first/last
grid-row which has content.

Bug: 885175
Change-Id: I563ad4ff19fba852c5526f88563eda2e430b7833
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3914454
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Kurt Catti-Schmidt <kschmi@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1052628}
NOKEYCHECK=True
GitOrigin-RevId: 660cc1d6fbd6202ebc2ccb6ba8e18d8f61e57574