Remove the skia/dawn code path for 2d canvas
We are no longer pursuing rendering 2d canvas content using dawn
directly in blink.  The new strategy is to favor the OOPR code path.
If we ever do end-up using dawn, it will not be in blink but rather on
the service side of the raster interface (in the GPU process).

Bug: 1355926
Change-Id: I0ba638912f5b34740cce7c1f786b8fb347e4477e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3851055
Commit-Queue: Justin Novosad <junov@chromium.org>
Reviewed-by: Juanmi Huertas <juanmihd@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1052704}
NOKEYCHECK=True
GitOrigin-RevId: ec25dd0505a12d629587814682f61f025bd4e430