[LazyEmbeds] Append iframe per each just before assertions in WPT
This CL addresses the flakiness of LazyEmbeds WPT. In the previous
CL[1], we secured the distance threshold to make sure the loading being
triggered by Intersection Observer rather than the time-based timeout.
But actually it doesn't solve the issue and we still see some test
faulures[2], specifically the assertion in L119[3]. It tests if the
iframe src is still not loaded before the scrolling (it expects false
but actually gets true sometimes).

Currently the target promise_test inserts 4 iframes at the same time[4],
and check if the frame src is loadeed or not loaded one by one, by
listening the each `load` event. The assertion which we'd like to fix is
located at the third position in the test assertions, that means the
assertion needs to wait for 2 iframes loading completions. My hypothesis
is that the target iframe triggers the loading based on the given
timeout while waiting for the previous iframe loading completions
depending on network conditions in test environments.

As a solution to that, this CL applies following changes: 1) Split the
current promise_test into two tests. The test for above the fold and the
other one is for below the fold. 2) Insert iframes just before the
assertion per each in order to avoid time consuming by waiting other
iframe loading. 3) Small refactoring to keep the scope small and improve
the readability.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/3874542
[2]
https://ci.chromium.org/ui/p/chromium/builders/ci/Linux%20Tests%20(dbg)(1)/108141/overview?q=ExactID%3Aninja%3A%2F%2F%3Ablink_wpt_tests%2Fvirtual%2Fautomatic-lazy-frame-loading%2Fwpt_internal%2Flazyembed%2Fautomatic-lazy-frame-loading.sub.html+VHash%3A9f5ebdc6f73354e7
[3]
https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/web_tests/wpt_internal/lazyembed/automatic-lazy-frame-loading.sub.html;l=119?q=automatic-lazy-frame-loading.sub.html
[4]
https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/web_tests/wpt_internal/lazyembed/automatic-lazy-frame-loading.sub.html;l=102?q=automatic-lazy-frame-loading.sub.html

Bug: 1359138
Change-Id: Ica2bd6fd617baa4554d9de7f1627386b4f28656e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3925343
Reviewed-by: Minoru Chikamune <chikamune@chromium.org>
Commit-Queue: Shunya Shishido <sisidovski@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1053399}
NOKEYCHECK=True
GitOrigin-RevId: 2606ff60a499e0362887d9b32932dee68913dfac