[GridNG] Rename is_parent_grid_container to is_subgridded_to_parent
Language of the variable is a bit misleading: subgrid line names may not
be applied even if the parent is a grid container, e.g., when the parent
grid of a subgrid has size containment.

This change shouldn't have any behavioral impact.

Bug: 618969
Change-Id: I5295a88e52ed734e1ade339027f45df6be451f7e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3923731
Commit-Queue: Ethan Jimenez <ethavar@microsoft.com>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1053435}
NOKEYCHECK=True
GitOrigin-RevId: 48fabbc9909bb4f12292aa1d1266bae59a1167a5