Transform timeouts in errors in COOP iframe WPTs.
When we got an unexpected no-opener behavior, we skipped verifying that
the popup handle was not null. This resulted in a javascript syntax
error (property access into null) and a timeout. Add an extra check
before so that we simply fail a check instead.

Bug: 1318373
Change-Id: Idd0f309d692c205c96a30e28467a13a8c42edf60
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3929045
Commit-Queue: Arthur Hemery <ahemery@chromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1053516}
NOKEYCHECK=True
GitOrigin-RevId: 47f4321f83ed1fa0a753f6218a5f8da66a0a7806