[scroll-animations] Include view-timeline properties in ComputedStyle diff
This is required for the style recalc machinery to understand that changes
in view-timeline properties may affect descendants and subsequent
siblings. This problem was already solved for scroll-timeline, so it's
a simple matter of also including view-timeline properties in the same
place.

Bug: 1317765
Change-Id: I93af47ac4b2ef6714edbbe0e8b6ce9f002c72485
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3898994
Reviewed-by: Kevin Ellis <kevers@chromium.org>
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1053606}
NOKEYCHECK=True
GitOrigin-RevId: 315e55b027ab95ef8fb75c36a876fefd9c5a2ce0