Replace ViewportToScreen with LocalRootToScreen
Replace ChromeClient's ViewportToScreen with LocalRootToScreen.
Requiring clients to convert to viewport coordinates (which is ambiguous
in an OOPIF as to what is meant) is subtle and varies depending on
whether called from an OOPIF or main frame because "viewport" is an
outermost frame concept. LocalRootToScreen is clearer and can
automatically perform the viewport conversion correctly for callers.

This is in preparation to a followup CL which will update
Element::VisibleBoundsInVisualViewport to be in local root coordinates
(as it currently is within an OOPIF) in all cases.

Note: ViewportToScreen (and now LocalRootToScreen) will not necessarily
return the screen coordinates of the given rect if called on a local
root while there is a visual viewport transform in the outermost widget.
That's because the widget screen rects don't account for the visual
viewport correctly. A previous attempt at this CL attempted to fix this
but was reverted: https://crrev.com/c/3891937. This CL leaves the bug in
to unblock changing Element::VisibleBoundsInVisualViewport.

Bug: 1314858
Change-Id: Ic589da8857752e4ecf638c1f1569264bc8d229ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3924242
Commit-Queue: David Bokan <bokan@chromium.org>
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1053695}
NOKEYCHECK=True
GitOrigin-RevId: 84313ca350fb82c8831e1e4abc5648b2523a808c