compute pressure: Implement pressure factor in Blink
As described in [1], pressureRecords contains also pressureFactors.
This patch is introducing the pressureFactors from the pressureManager.

The algorithm to extract the pressure factor from the platformCollector
is not yet implemented and therefore pressureFactor value is hardcoded
in this patch.

The Blink part is fully functional and tested with new wpt tests.

[1] https://wicg.github.io/compute-pressure/#contributing-factors

Bug: 1356191
Change-Id: I0e4b6c32244729d92108f4b58c573dc6df0cdbfb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3904313
Reviewed-by: Fr <beaufort.francois@gmail.com>
Reviewed-by: Raphael Kubo Da Costa <raphael.kubo.da.costa@intel.com>
Reviewed-by: Will Harris <wfh@chromium.org>
Commit-Queue: Arnaud Mandy <arnaud.mandy@intel.com>
Cr-Commit-Position: refs/heads/main@{#1054113}
NOKEYCHECK=True
GitOrigin-RevId: 6924d07f66c1f7c65da55b6e3c9b2070992fd1e4