Replace SVGElement in the SVGResources interface with LayoutObject
This better reflects the actual intended use of this interface - not
what the resource client is owned by (hopefully that will change in the
future).

Drop redundant arguments where possible to clean up callsites. The
<paint> and <marker> functions still have to accept "redundant"
arguments to some degree because the Update*() function uses the
Clear*() function. For <paint> we also need to be able to pass an
independent ComputedStyle because of updates/clears (for ::selection).

Bug: 1274314
Change-Id: Iee601212c410e812dddc5c1efca0392f490af710
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3930751
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Auto-Submit: Fredrik Söderquist <fs@opera.com>
Commit-Queue: Stephen Chenney <schenney@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1054195}
NOKEYCHECK=True
GitOrigin-RevId: c2e15bcdac673b1bb8c1acdef037ef73df964d32