Support layout caching of fragmented elements.
Even if an element breaks into multiple fragments, there are cases where
we can re-use the fragments from the cache, as long as the fragmentainer
block-size is the same, and the offset to the fragmentation line is the
same as before, and there are no column spanners involved.

To support this, LayoutBox::GetCachedLayoutResult() now takes a break
token parameter. Call sites that know that there's only one fragment
should call the new function GetSingleCachedLayoutResult(), to assert
that no fragmentation is involved.

This avoids O(n^2) performance complexity in nested tables that are
fragmented, and makes
perf_tests/layout/multicol/deeply-nested-tables.html about 1250 times
faster.

Bug: 1353238
Change-Id: Id8644b1a984f648991a2f1c7cd461bfea451e0d3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3925615
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1054205}
NOKEYCHECK=True
GitOrigin-RevId: b52b7957e88022257ac515866664f0bd3172233c