Clone fragment items as part of child cloning.
A fixed-positioned repeated element may have inline children, and, since
the entry-point into the fragment repeater is CloneChildFragments(), we
need to handle it there. It used to be in Repeat(), which was an
acceptable home as long as we only supported repetition of table
sections (they never have inline children), but for fixed-positioned
elements this matters.

Bug: 1370305
Change-Id: I9edd8d65f3d24f5b3a9f0c28af3ce434d5468d85
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3930752
Auto-Submit: Morten Stenshorne <mstensho@chromium.org>
Reviewed-by: Alison Maher <almaher@microsoft.com>
Commit-Queue: Alison Maher <almaher@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1054439}
NOKEYCHECK=True
GitOrigin-RevId: 1f815457368e018fca3154ada109f3f53110cbff