Avoid sending FP and FCP UKM on soft navigations
In [1], we added support for FP and FCP in soft navigations, but did not
properly turn off their UKM reporting in those cases (as we're not yet
reporting soft navigation page load metrics).

This CL fixes that, by turning off the reporting, as well as overriding
the reported values to ensure the soft navigation generated values are
not reported when other timing info is.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/3913570

Bug: 1370495
Change-Id: I6760af558d031a021693e0368bf0e9c757e294d4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3928699
Commit-Queue: Yoav Weiss <yoavweiss@chromium.org>
Reviewed-by: Ian Clelland <iclelland@chromium.org>
Reviewed-by: Lan Wei <lanwei@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1054764}
NOKEYCHECK=True
GitOrigin-RevId: cb0928d77d6ac1f0522418f6668c4a75ba31fc30