Add microtask queues and rejected promises per WindowAgent.
Allow for an agent to query the WebAgentGroupScheduler. This will be
useful in the future as a detached window is associated with an agent
and we will be able to get the WebAgentGroupScheduler.

Associate an Agent with the WebAgentGroupScheduler. This allows us to
execute the microtask queues for each Agent.

RejectPromises per window agent is enabled by default.
Microtaskqueue per window agent is still being worked on there are
still a number things to address before it can be enabled.

BUG=961186

Change-Id: I1294580dc4a2d86fe40950325d46a51b8c9dd64e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3928076
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Dave Tapuska <dtapuska@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1055420}
NOKEYCHECK=True
GitOrigin-RevId: 1f8ed4479f863fa19de009fcff617795516ab7c7