Add grpc-java to 3pp.
Does not update DEPS/BUILD.gn because the packages are not in CIPD yet.

I'll wait until the children CL is approved to submit this one, then I'll submit this one and wait for the packages to be in CIPD, then I'll check that the child CL works (gclient sync and build) and submit it.

Change-Id: Icab6db4ae61e52671bff5b6edfde18fa7486720a
Bug: b:245087054
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3854326
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Commit-Queue: David Bengoa <dbengoa@google.com>
Cr-Commit-Position: refs/heads/main@{#1056183}
NOKEYCHECK=True
GitOrigin-RevId: 37fa59673f64560670698238770b367abdbba065