Perf test for balanced multicol with tall list item markers.
If inside a balanced multicol container, we have a performance problem
if list item markers are just a little bit taller than the line that
they are next to. The problem is that this provides a very small
NGLayoutResult::MinimalSpaceShortage(), so that we'll end up performing
very many layout passes, stretching the columns just a tiny bit at a
time. The test has no perfect break opportunities, so the end result is
that everything fits in one column.

Bug: 1371575
Change-Id: I79a6be884c73ea349e3c39f39d07bd07b374df51
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3934286
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1056437}
NOKEYCHECK=True
GitOrigin-RevId: c573d5f08742a68629aa8eb7dd79555ddbede2c0