Move ScrollTimeline post-layout validation to LayoutClean
ScrollTimeline post-layout validation requires only clean style and
layout, as it only needs the dimensions and scroll offsets of the
scroll containers.

This patch moves it to LayoutClean so that anchor positioning can share
the same revalidation timing. In particular, we want to pass the
validity of the snapshot to the paint property tree, so we must do
revalidation before PrePaint.

This patch also catches a minor issue that the document may be left
with NeedsLayoutTreeUpdate after a full lifecycle update:
- ScrollTimeline::ValidateState() updates timeline_state_snapshotted_,
  which marks the timeline dirty, and hence NeedsLayoutTreeUpdate for
  the document
- The timeline doesn't have any connected effect target element, so
  no element is marked style dirty, and hence lifecycle isn't rewinded
  and we don't repeat style and layout for it

Bug: 1371217
Change-Id: I745e442b44f85111dd7367590b45c1d1851ec6ac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3938621
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Commit-Queue: Xiaocheng Hu <xiaochengh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1056598}
NOKEYCHECK=True
GitOrigin-RevId: 7b6bf8870ecd71d5514aa9d31a229e9a271086cc