[ThreadedBodyLoader Part 1] Handle decoded data in DocumentLoader
This is the first change to enable reading and decoding the navigation
body earlier in the commit process and off the main thread. For more
info and motivation, see the bug.

This change modifies DocumentLoader and related classes to allow
receiving and processing decoded data. As much as possible, we try to
use the same code path for decoded or encoded data to avoid forking any
of the logic.

Future patches will add support to NavigationBodyLoader to load and
decode the body in a background thread, and then call
DecodedBodyDataReceived() with the result.

A prototype of this optimization has shown good results for some sites
on Pinpoint. Using the Google_cold story in loading.desktop:

Windows: https://pinpoint-dot-chromeperf.appspot.com/job/15f1a62b1a0000
-12.1% FCP, -4.6% LCP

Linux: https://pinpoint-dot-chromeperf.appspot.com/job/127cd5671a0000
-12.9% FCP, -1.5% LCP

Bug: 1373213
Change-Id: I09393f239349627c6de31c267e080a2e2e7e70aa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3943290
Reviewed-by: Nate Chapin <japhet@chromium.org>
Commit-Queue: Clark DuVall <cduvall@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1057245}
NOKEYCHECK=True
GitOrigin-RevId: f6445d5a876f4b0adf110e3fec6db661c8c02313