WebNN: validate the dimensions before create output operand
This CL implements the validation of output operand for operation build
methods. The MLOperand::ValidateAndCreateOutput() method enforces the
validation of elements number and byte length of dimensions before
creating the operand. This CL also moves the input and constant
validation to MLOperand::ValidateAndCreateInput() and
MLOperand::ValidateAndCreateConstant() respectively.

This CL also fixes an issue of not using checked math to calculate the
byte length. To facilitate the graph builder, such as reshape operation,
and the backend implementation, the MLOperand::NumberOfElements() and
MLOperand::ByteLength() are added for better code reuse.

This CL also adds unit tests to cover the output operand validation.

Bug: 1273291
Change-Id: I2520c28f0e96ab57462969b8fc1a75cef3914120
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3929172
Commit-Queue: ningxin hu <ningxin.hu@intel.com>
Reviewed-by: Jiewei Qian <qjw@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1057320}
NOKEYCHECK=True
GitOrigin-RevId: 68ca6a222b1c70a002005c247ee6b3063b5f38ca