[Cast Streaming] Fallback to mirroring on navigation
Currently, when remoting if the back or forward navigation buttons
are pressed the remoting session freezes instead of falling back to
mirroring.

Turns out, this is due to the fact that the page doesn't necessarily
get destroyed on navigation, and the HTMLMediaElement and its web player
may stick around, just in a paused state.

Obviously we don't want to stop remoting if the video is paused, so the
solution in this patch is to allow the media player to know when the
page is frozen and communicate that to its observer.

Bug: 1147662
Change-Id: I94a024548486aca81afe5bb945ff32c6577ddc63
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3936590
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Mark Foltz <mfoltz@chromium.org>
Commit-Queue: Jordan Bayles <jophba@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1057328}
NOKEYCHECK=True
GitOrigin-RevId: db0fc976fe1b11f7a09a409deb92691549450f07