[GridNG] Build up NGGridSizingTree for subgrid layout
In order to correctly perform subgrid layout we need to do a significant
refactor in NGGrid to allow the topmost grid container to size the whole
grid tree. This is required since we need to provide the contribution
size of every subgrid up to the same step in the parent grid's sizing
algorithm, which we are not able to easily do in `ComputeMinMaxSizes`
nor `Layout` because it expects grid sizing to be finalized.

This CL introduces `NGGridSizingTree`, a class representing a tree of
multiple nested grids, with the root of the tree being the topmost grid
and the rest of the nodes being subgrids (at least one subgridded axis).

Our tree structure is "flattened", meaning it's contained in a single
vector with each tree node storing a pointer to its parent (except the
position 0, which is the root) and built in such a way that the nodes of
every subtree are contiguous in the vector. To perform grid sizing, each
node stores data for their respective grid's algorithm, e.g., its layout
track collections, grid items, placement data, etc.

Bug: 618969
Change-Id: I8052b0e8a41cd9207f46597c4bbe620f26a190d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3940365
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: Ethan Jimenez <ethavar@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1057396}
NOKEYCHECK=True
GitOrigin-RevId: c4d71ae4f9f05d10d4905f0e27a793f1a3b7d19e