[last-baseline] Fix inconsistent wm+direction within table internals.
This moves the style adjuster fixup for tables into the layout-tree
builder.

Previously upon the body to root writing mode propagation it was
possible for the style of table internals to be inconsistent. By
moving the fixup into the tree-builder we can ensure it is correct,
and has the side effect of reporting the correct style from computed
style.

Fixed: 1366340, 736093
Change-Id: I261010f8eb6de051a4a8ae112303064c9e426a15
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3938617
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1057431}
NOKEYCHECK=True
GitOrigin-RevId: 02390308d123c0b8f90934489eee3f19ca596ec5