[webdriver] Add Get All/Named Cookie Command {Fix SameSite}
In the prior CL there was a request to treat specified Lax and default
Lax different as they are treated different in chrome:
https://chromium-review.googlesource.com/c/chromium/src/+/3891100/comment/075cdb5a_5770c84b/
Testing on the WPT side (cross browser) has informed me that the current
standard is not to differentiate. As the difference here is chrome
specific, it seems reasonable to just return Lax.
https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Set-Cookie/SameSite

Bug: 1342633
Change-Id: I0ff0a99b704a69be14eaf3847456aad1eecc877b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3945458
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Auto-Submit: Ari Chivukula <arichiv@chromium.org>
Commit-Queue: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1057744}
NOKEYCHECK=True
GitOrigin-RevId: 393cd868fdc0c496b3b67c4182788134b52c2435