[webdriver] Add Get All/Named Cookie Command {Fix Rejection}
It turns out that when we try to get a specific named cookie and we
cannot do so because the cookie does not exist, we need to throw an
error and not just return null. Let's handle this in the js layer for
simplicity.
https://w3c.github.io/webdriver/#get-named-cookie

Bug: 1342633
Change-Id: Idff0eeda473c3757d9864c51fa97f1aab0467474
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3945601
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Auto-Submit: Ari Chivukula <arichiv@chromium.org>
Commit-Queue: Ari Chivukula <arichiv@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1057791}
NOKEYCHECK=True
GitOrigin-RevId: a44c9b5cc02990f3012105b77245766e2532664c