SET: Improve viewport documentation and naming
The CL introduces the concept of the "Snapshot Viewport", adds
documentation, renames things with this in mind, and makes the code
clearer. This CL has no intended behavioral changes.

The main change is to replace the "ViewportWidgetInsets" and
"RootContainerSize" methods with GetSnapshotViewportRect, which includes
the offset from the fixed viewport, and GetRootSnapshotPaintOffset,
which is used to offset painting of the root snapshot. The "widget
insets" are still used internally to compute these values.

Also rename "viewport_matrix" to "snapshot_matrix" since this is the
transform from local coordinates into the snapshot viewport.

Bug: 1357156
Change-Id: I529324336c762b4f1dbe086f4aa0afaa440f4045
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3938846
Commit-Queue: David Bokan <bokan@chromium.org>
Reviewed-by: Khushal Sagar <khushalsagar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1057830}
NOKEYCHECK=True
GitOrigin-RevId: 21288e1ca8bed16d128aa84dc965f8efb0e7cd2f